ContenidoAlumniApi/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
daphne==4.2.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
Markdown==3.10.2
msgpack==1.1.2
pillow==12.1.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
redis==7.3.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
